This PR improves performance of the CXXRTL simulations by approximately 2.6× uniformly across different pass pipelines on a single representative design (Minerva SRAM SoC with black box UART). Moreover, if the design is fully flattened, it is able to achieve 1 delta cycle per clock edge by optimally scheduling all of the logic in the design, including the black box.

The following is a totally non-scientific estimate of the speedup provided by this PR. It was collected on an idle Dell XPS13 9360 laptop with Intel Core i7-7500U CPU, Linux 4.19.0-8, and clang++ 7.0.1-8 by running the Minerva SRAM SoC for 1 million cycles. Each evaluation step invokes the `poll()` syscall (and very few of them also `write()`), which are excluded from the measurements (i.e. only user time is counted). [Reproduce my results!](https://github.com/YosysHQ/yosys/files/4516222/cxxrtl-toymark.zip)

These numbers are not intended to quantify any specific aspect of CXXRTL (much less serve as a comparison with other tools), but rather give an approximate sense of the kind of performance one might expect when simulating a small but practical SoC.

| passes          | before   | after    | speedup |
|-----------------|----------|----------|---------|
|  none           | 22 kCPS  | 55 kCPS  |  +150%  |
| `proc`          | 32 kCPS  | 90 kCPS  |  +180%  |
| `proc; flatten` | 280 kCPS | 730 kCPS |  +160%  |

I would also like to thank @jfng for his contribution to this work.